Remove rejections for QMessageFilter functions, add LikeComparator enum
byCc, byTo and byBcc rejects were removed and the LikeComparator enum
was added to the Messaging module. The version of byRecipients that
uses LikeComparator is working now, but the other two versions are
still being rejected because of the same problem: there are declarations
of them in the header but no implementation.